pyproject.toml
[tool.poetry]
name = "crbnet-works-cdk"
version = "0.1.0"
description = ""
authors = ["Chris Bowles <BowlesCR@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
aws-cdk-lib = "^2.96.2"
constructs = "^10.2.70"


[tool.poetry.group.dev.dependencies]
black = "^23.9.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 160
target-version = ['py311']